requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
ultralytics
streamlit
Pillow
opencv-python-headless
numpy